vim-patch:8.2.0443: clipboard code is spread out
Problem:    Clipboard code is spread out.
Solution:   Move clipboard code to its own file. (Yegappan Lakshmanan,
            closes vim/vim#5827)

https://github.com/vim/vim/commit/45fffdf10b7cb6e59794e76e9b8a2930fcb4b192
