+++ src/nvim/clipboard.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <stdbool.h>
+
+#include "nvim/register_defs.h"
+
+#include "clipboard.h.generated.h"
